% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_summary_table.R
\name{add_font_for_significant}
\alias{add_font_for_significant}
\title{Add background for significant values}
\usage{
add_font_for_significant(wb, sheet, table)
}
\arguments{
\item{wb}{A Workbook object from xlsx library}

\item{sheet}{A sheet object from Workbook}

\item{table}{A data.frame containing the data to be saved, i.e. summary_table
object}
}
\description{
This function enables to add lavender background for cells
corresponding to a significant p value
}
