% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_summary_table.R
\docType{data}
\name{default_columns}
\alias{default_columns}
\title{Function to save summary table in a nicely formatted excel file}
\format{
An object of class \code{list} of length 2.
}
\usage{
default_columns
}
\arguments{
\item{summary_table}{A data.frame containing the data to save in excel file}

\item{filepath}{A character containing a path to the folder for saving}

\item{filename}{A character containing the name to give to the file
(without the extension)}

\item{sheetname}{A character containing the name to give to the excel sheet}

\item{title}{A character containing a title to put as header in the sheet.
Default :""}

\item{subtitle}{A character containing a subtitle appearing rigth below the
title. Default :""}

\item{append}{Boolean. If TRUE, save_summary_table will open an existing
file. If FALSE (default), it creates a new file and erase already existing
file with same name.}

\item{list_variables_renaming}{A list whose names correspond to variables to
rename and values to the new names to use}

\item{list_columns_renaming}{A list whose names correspond to columns to
rename and values to the new names to use}

\item{underline_p}{A boolean, indicating whether to add a font or not, to
significant pvalues (default: True)}
}
\value{
Nothing. But an excel file is saved at the given location.
}
\description{
Function to save summary table in a nicely formatted excel file.
It can use an already existing excel file and add a sheet. Or, it can create
a new file.
}
\examples{
\dontrun{
# Creating table to save
table <-summary_table(data = wdbc.data,
                      studied_vars = c("radius", "texture", "compactness_quartile"),
                      dependent = "diagnosis",
                      multivariate = list(c("smoothness", "texture"),
                                          c("concavity", "symmetry"))
                                          )
# Creating an excel file
save_summary_table(table, filepath="data/", filename="test",
                   sheetname = "test_sheet_1",
                   title = "Regression logistic study",
                   subtitle = "This is an interesting study.",
                   list_variables_renaming = list("compactness_quartile"="Compactness quartile")
                   )

# Adding a sheet to an existing excel file
save_summary_table(table, filepath="data/", filename="test",
                   sheetname = "test_sheet_2",
                   title = "Regression logistic study",
                   subtitle = "This is an interesting study.", append=TRUE)
}

}
\keyword{datasets}
