% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_summary_table.R
\name{extract_pvalue_from_char}
\alias{extract_pvalue_from_char}
\title{Extracting pvalue from string}
\usage{
extract_pvalue_from_char(char_variable)
}
\arguments{
\item{char_variable}{A character from which to extract p value}
}
\value{
A numeric value corresponding to pvalue
}
\description{
Extracting pvalue from string
}
