% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_summary_table.R
\name{set_custom_style}
\alias{set_custom_style}
\title{Creating a cell style}
\usage{
set_custom_style(wb, list_args)
}
\arguments{
\item{wb}{A workbook object}

\item{list_args}{A list containing differents lists of arguments : list_font,
list_alignment, and list_border. Those sublists contain respectively
arguments for Font, Alignment, and Border functions of CellStyle.}
}
\value{
A CellStyle defined thanks to \code{list_args}
}
\description{
A function to define an xlsx cell style, depending on a list of
arguments \code{list_args}
}
