% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_transform_OR_with_IC.R
\name{str_transform_OR_with_IC}
\alias{str_transform_OR_with_IC}
\title{Title}
\usage{
str_transform_OR_with_IC(OR, IC_min, IC_max, pvalue, digits = 3, digits_p = 3)
}
\arguments{
\item{OR}{A numeric input corresponding to an Odd Ratio}

\item{IC_min}{A numeric input corresponding to the lower boundary of the 95\%
interval associated to OR}

\item{IC_max}{A numeric input corresponding to higher boundary of the 95\%
interval associated to OR}

\item{pvalue}{A numeric input corresponding to the pvalue associated to OR}

\item{digits}{An integer input giving the number of digits for rounding OR
and IC}

\item{digits_p}{An integer input giving the number of digits for rounding
p_value}
}
\value{
a string "OR (IC_min-IC_max, p=pvalue)"
}
\description{
Function to summarize OR, IC, pvalue from their numerical values
in a string of the form "OR (IC_min-ICmax, p=pvalue)"
}
\examples{
str_transform_OR_with_IC(OR=2.02, IC_min=1.803, IC_max=2.405, pvalue=0.0025)
str_transform_OR_with_IC(OR=2.02, IC_min=1.803, IC_max=2.405, pvalue=0.0025,
digits=1, digits_p=4)

}
