% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Title}
\usage{
summary_table(
  data,
  studied_vars,
  dependent,
  univariate = TRUE,
  multivariate = NULL
)
}
\arguments{
\item{multivariate}{}
}
\description{
Title
}
\examples{
summary_table(data = wdbc.data,
              studied_vars = c("radius", "texture"),
              dependent = "diagnosis",
              univariate = FALSE
             )

summary_table(data = wdbc.data,
              studied_vars = c("radius", "texture"),
              dependent = "diagnosis" )

summary_table(data = wdbc.data,
              studied_vars = c("radius", "texture"),
              dependent = "diagnosis",
              multivariate = list(c("smoothness", "texture"),
                                  c("concavity", "symetry"))
             )
}
