% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_summary_table.R
\name{xlsx.addTitle}
\alias{xlsx.addTitle}
\title{Function to add a title to excel file}
\usage{
xlsx.addTitle(sheet, rowIndex, title, titleStyle)
}
\arguments{
\item{sheet}{A sheet object in which title will be written}

\item{rowIndex}{A numeric value indicating the row to contain the title}

\item{title}{A character to be used as title}

\item{titleStyle}{A cell style object to use for the title}
}
\description{
This functions is an helper to build an excel output. It enables
to add a title, or a subtitle, to a sheet object from a workbook (using xlsx
library)
}
