% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_OR_to_str.R
\name{extract_OR_to_str}
\alias{extract_OR_to_str}
\title{Compute and summarize in a string OR, IC, pvalue for a glm model}
\usage{
extract_OR_to_str(model, studied_var, level = NULL)
}
\arguments{
\item{model}{A glm model corresponding to a logistic regression
(family=binomial)}

\item{studied_var}{Name of the variable to explain}

\item{level}{A string indicating the level to consider for categorical
variables (default:NULL)}
}
\value{
A string of the form "OR (IC_min-IC_max, p=pvalue)
}
\description{
Compute values for OR, IC, pvalue for a glm model and summarizes
result in a string of the form "OR (IC_min-IC_max, p=pvalue)"
}
\details{
Title
}
\examples{
model_1 <- glm( diagnosis ~ texture + radius + compactness_quartile,
                data=wdbc.data,
                family="binomial")

# For quantitative variable
extract_OR_to_str(model_1, studied_var = "texture")

# For categorical variable
extract_OR_to_str(model_1, studied_var = "compactness_quartile", level="2")
}
