% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table_utils.R
\name{extract_OR_to_table}
\alias{extract_OR_to_table}
\title{Function to complete OR part of summary table for a given quantitative or
categorical variable}
\usage{
extract_OR_to_table(data, table, studied_var, model, OR_colname)
}
\arguments{
\item{data}{A data.frame containing the data for Logistic Regression}

\item{table}{A data.fram containing the summary table}

\item{studied_var}{A character containing name of the variable of interest}

\item{model}{GLM model to use for OR extraction}

\item{OR_colname}{A character containing name of the final column}
}
\value{
table data.frame completed
}
\description{
Function to complete OR part of summary table for a given quantitative or
categorical variable
}
\keyword{internal}
